third_party/blink/web_tests/media/media-fragments/TC0012-expected.txt


Title: TC0012
Fragment: 't=3,'
Comment: Invalid syntax, hence the temporal fragment is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
